third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-status-code-states-expected.txt

[Worker] Test EventSource states for different status codes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-status-code-states.js
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 200 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 204 resulted in states CONNECTING, CLOSED, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 205 resulted in states CONNECTING, CLOSED, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 202 resulted in states CONNECTING, CLOSED, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 301 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 302 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 303 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 307 resulted in states CONNECTING, OPEN, OPEN, CONNECTING, CLOSED
PASS [Worker] arrayCompare(states, tests[count].expectedStates) is true
PASS [Worker] status code 404 resulted in states CONNECTING, CLOSED, CLOSED
[Worker] DONE
PASS successfullyParsed is true

TEST COMPLETE
